Separate data transformation from report generation
For HTML reports we need to have an GET endpoint instead of a post so
that a report can be shown via a browser without the need for the actual
report data.

To support that this commit separates the data transformation by adding
a post endpoint and storing the transformed json.
This endpoint returns an identifier which then can be used to generate
the actual report as an path parameter on `/report/$ID`.

To support visits of an link more than once a cache will be introduces to
prevent unnecessary generation of a report.