adds a workaround for inline svg and WeasyPrint
Unfortunately WeasyPrint does not support inline svg at all, since we
need that functionality to create pictures dynamically based on
report-formats.

This workaround does look for tags with svg and then creates an img
element with base64 image content of that.

We are painfully aware that we lose css information with that method,
but the conclusion is that it is better to may define design element twice
(in the svg and in css) than to not have this possibility at all.