create templatetags for charts instead of creating it in gvmd
To be capable of change the charts more transparently for templates the
chart creation moved from within the datastructure as data links to
within the template as svg via custom tags.

The data changes therefore as well: from containing the actual charts to
just providing the data in a more or less standardized fashion.

This commit contains the custom_tags:

- pie_chart
- h_bar_chart

those can be used within a template by

{{ overview.nvts | pie_chart }}
{{ overview.hosts | h_bar_chart }}