fix just the first svg will be replaced, fix bar chart
The bar chart is missing an closing g element and the svg image
replacement worked just for the first svg and will then produce
gibberish since the index has changed on html change.

In the moment it is using recursion, which should be replaced because
python does not nor will ever optimize tail recursion.