Extend bar chart with x titles, add bar chart as tag
Add possibility to add a title to the count numbers on the x-axis

To use the whole functionality of bar_h without having to specify every
parameter in order it is also registered as a simpletag instead of just
a filter.

This allows templates to use the chart by calling it:

```
{%  h_bar_chart overview.hosts x_title="Vulnerabilitites" %}
```

In addition to that there is also a new field called pheme_version with
just the numbers of set version.

This allows feature_toggles within the templates e.g.:

```
{% if pheme_version and pheme_version > 2144 %}
{%  h_bar_chart overview.hosts x_title="Vulnerabilitites" %}
{% else %}
{{ overview.hosts |  h_bar_chart }}
{% end %}
```